requirements.txt
pillow
thumbhash-python
tyro
